refactor(refs T30969): move DpTimePicker to components root
