+++ src/components/DpTimePicker/DpTimePicker.vue
@@ -87,12 +87,12 @@ export default {
 
   components: {
     DpInput: async () => {
-      return await import('../../DpInput/DpInput')
+      return await import('../DpInput/DpInput')
     },
     DpLabel: async () => {
-      return await import('../../DpLabel/DpLabel')
+      return await import('../DpLabel/DpLabel')
     },
-    DpResettableInput: () => import('../DpResettableInput')
+    DpResettableInput: () => import('../core/DpResettableInput')
   },
 
   directives: {
